% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeter.R
\name{greeter}
\alias{greeter}
\title{rsalamz}
\usage{
greeter(name)
}
\arguments{
\item{name}{This is the name.}
}
\value{
statement that greets people in the morning.
}
\description{
A function that is used to greet people in the morning.
}
\examples{
greeter("Patricia")
}
